Fix after review, WIP.

+++ src/Elrond/GasEstimator.cpp
@@ -12,26 +12,26 @@ using namespace TW;
 using namespace TW::Elrond;
 
 // Additional gas to account for eventual increases in gas requirements (thus avoid breaking changes in clients of TW-core).
-const uint32_t ADDITIONAL_GAS_FOR_ESDT_TRANSFER = 100000;
+const uint64_t ADDITIONAL_GAS_FOR_ESDT_TRANSFER = 100000;
 
 // Additional gas to account for extra blockchain operations (e.g. data movement (between accounts) for NFTs), 
 // and for eventual increases in gas requirements (thus avoid breaking changes in clients of TW-core).
-const uint32_t ADDITIONAL_GAS_FOR_ESDT_NFT_TRANSFER = 500000;
+const uint64_t ADDITIONAL_GAS_FOR_ESDT_NFT_TRANSFER = 500000;
 
 GasEstimator::GasEstimator(const NetworkConfig& networkConfig) {
     this->networkConfig = networkConfig;
 }
 
-uint32_t GasEstimator::forEGLDTransfer(uint32_t dataLength) {
-    uint32_t gasLimit = 
+uint64_t GasEstimator::forEGLDTransfer(size_t dataLength) {
+    uint64_t gasLimit = 
         this->networkConfig.getMinGasLimit() + 
         this->networkConfig.getGasPerDataByte() * dataLength;
 
     return gasLimit;
 }
 
-uint32_t GasEstimator::forESDTTransfer(uint32_t dataLength) {
-    uint32_t gasLimit = 
+uint64_t GasEstimator::forESDTTransfer(size_t dataLength) {
+    uint64_t gasLimit = 
         this->networkConfig.getMinGasLimit() + 
         this->networkConfig.getGasCostESDTTransfer() + 
         this->networkConfig.getGasPerDataByte() * dataLength +
@@ -40,8 +40,8 @@ uint32_t GasEstimator::forESDTTransfer(uint32_t dataLength) {
     return gasLimit;
 }
 
-uint32_t GasEstimator::forESDTNFTTransfer(uint32_t dataLength) {
-    uint32_t gasLimit = 
+uint64_t GasEstimator::forESDTNFTTransfer(size_t dataLength) {
+    uint64_t gasLimit = 
         this->networkConfig.getMinGasLimit() + 
         this->networkConfig.getGasCostESDTNFTTransfer() + 
         this->networkConfig.getGasPerDataByte() * dataLength +
